authentication/app/requirements.txt
asgiref==3.8.1
attrs==24.2.0
Authlib==1.3.2
autobahn==24.4.2
Automat==24.8.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==43.0.3
Django==4.2.16
django-authlib==0.17.1
django-formtools==2.5.1
django-otp==1.5.4
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
oauthlib==3.2.2
packaging==24.2
password-strength==0.0.3.post2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.2
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
zope.interface==7.1.1
